+++ src/devuni.h
@@ -42,53 +42,53 @@ using namespace std;
 class DevUni: public Quality {
  public:
 
-  vector<long double> in; // used to compute the quality participation of each community
+  vector<float> in; // used to compute the quality participation of each community
   vector<int> w; // used to store size of communities
 
   DevUni(Graph & gr);
   ~DevUni();
 
-  inline void remove(int node, int comm, long double dnodecomm);
+  inline void remove(int node, int comm, float dnodecomm);
 
-  inline void insert(int node, int comm, long double dnodecomm);
+  inline void insert(int node, int comm, float dnodecomm);
 
-  inline long double gain(int node, int comm, long double dnodecomm, long double w_degree);
+  inline float gain(int node, int comm, float dnodecomm, float w_degree);
 
-  long double quality();
+  float quality();
 };
 
 
 inline void
-DevUni::remove(int node, int comm, long double dnodecomm) {
+DevUni::remove(int node, int comm, float dnodecomm) {
   assert(node>=0 && node<size);
 
-  in[comm] -= 2.0L*dnodecomm + g.nb_selfloops(node);
+  in[comm] -= 2.0*dnodecomm + g.nb_selfloops(node);
   w[comm]  -= g.nodes_w[node];
 
   n2c[node] = -1;
 }
 
 inline void
-DevUni::insert(int node, int comm, long double dnodecomm) {
+DevUni::insert(int node, int comm, float dnodecomm) {
   assert(node>=0 && node<size);
 
-  in[comm] += 2.0L*dnodecomm + g.nb_selfloops(node);
+  in[comm] += 2.0*dnodecomm + g.nb_selfloops(node);
   w[comm]  += g.nodes_w[node];
 
   n2c[node] = comm;
 }
 
-inline long double
-DevUni::gain(int node, int comm, long double dnc, long double degc) {
+inline float
+DevUni::gain(int node, int comm, float dnc, float degc) {
   assert(node>=0 && node<size);
 
-  long double wc = (long double)w[comm];
-  long double wu = (long double)g.nodes_w[node];
+  float wc = (float)w[comm];
+  float wu = (float)g.nodes_w[node];
   
-  long double m2 = g.total_weight;
-  long double n  = (long double)g.sum_nodes_w;
+  float m2 = g.total_weight;
+  float n  = (float)g.sum_nodes_w;
 
-  long double gain = dnc - (m2*wu*wc)/(n*n);
+  float gain = dnc - (m2*wu*wc)/(n*n);
 
   return gain;
 }
