+++ src/goldberg.h
@@ -42,60 +42,60 @@ using namespace std;
 class Goldberg: public Quality {
  public:
 
-  vector<long double> in; // used to compute the quality participation of each community
+  vector<float> in; // used to compute the quality participation of each community
   vector<int> w; // used to store size of communities
 
-  long double max; // biggest weight on links
+  float max; // biggest weight on links
 
-  Goldberg(Graph & gr, long double max_w);
+  Goldberg(Graph & gr, float max_w);
   ~Goldberg();
 
-  inline void remove(int node, int comm, long double dnodecomm);
+  inline void remove(int node, int comm, float dnodecomm);
 
-  inline void insert(int node, int comm, long double dnodecomm);
+  inline void insert(int node, int comm, float dnodecomm);
 
-  inline long double gain(int node, int comm, long double dnodecomm, long double w_degree);
+  inline float gain(int node, int comm, float dnodecomm, float w_degree);
 
-  long double quality();
+  float quality();
 };
 
 
 inline void
-Goldberg::remove(int node, int comm, long double dnodecomm) {
+Goldberg::remove(int node, int comm, float dnodecomm) {
   assert(node>=0 && node<size);
 
-  in[comm] -= 2.0L*dnodecomm + g.nb_selfloops(node);
+  in[comm] -= 2.0*dnodecomm + g.nb_selfloops(node);
   w[comm]  -= g.nodes_w[node];
 
   n2c[node] = -1;
 }
 
 inline void
-Goldberg::insert(int node, int comm, long double dnodecomm) {
+Goldberg::insert(int node, int comm, float dnodecomm) {
   assert(node>=0 && node<size);
 
-  in[comm] += 2.0L*dnodecomm + g.nb_selfloops(node);
+  in[comm] += 2.0*dnodecomm + g.nb_selfloops(node);
   w[comm]  += g.nodes_w[node];
 
   n2c[node] = comm;
 }
 
-inline long double
-Goldberg::gain(int node, int comm, long double dnc, long double degc) {
+inline float
+Goldberg::gain(int node, int comm, float dnc, float degc) {
   assert(node>=0 && node<size);
 
-  long double inc  = in[comm];
-  long double self = g.nb_selfloops(node);
-  long double wc   = (long double)w[comm];
-  long double wu   = (long double)g.nodes_w[node];
+  float inc  = in[comm];
+  float self = g.nb_selfloops(node);
+  float wc   = (float)w[comm];
+  float wu   = (float)g.nodes_w[node];
   
-  long double gain;
+  float gain;
 
-  if (wc == 0.0L)
-    gain  = (2.0L*dnc+self) / (2.0L*wu);
+  if (wc == 0.0)
+    gain  = (2.0*dnc+self) / (2.0*wu);
   else {
-    gain  = (2.0L*dnc+self + inc) / (2.0L*(wc+wu));
-    gain -= inc / (2.0L*wc);
+    gain  = (2.0*dnc+self + inc) / (2.0*(wc+wu));
+    gain -= inc / (2.0*wc);
   }
 
   return gain;
